+++ src/plugins/extra/widgets/index.ts
@@ -1,2 +1,3 @@
+export * from './quote';
 export * from './search';
 export * from './weather';