+++ src/plugins/extra/widgets/quote/index.ts
@@ -1,0 +1,2 @@
+export { default as Quote } from './Quote';
+export { default as QuoteSettings } from './QuoteSettings';