+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+- New Daily Quotes plugin to get your inspired (or laughing) each day.
+
 ## [1.9.0] - 2018-01-06
 ### Added
 - The first greeting translations in French and Korean! (and a few other dashboard items)